requirements.txt
ollama>=0.6.1
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
